web/requirements.txt
Django==3.2
psycopg2-binary==2.8
pytest==6.2.3
factory_boy==3.2.0
djangorestframework==3.12.4
pytest-django==4.2.0
requests==2.25.1
pytest-mock==3.6.0